+++ test/test_helper.rb
@@ -9,7 +9,6 @@ Reform::Form.class_eval do
 end
 
 require "trailblazer/operation/builder"
-require "trailblazer/operation/setup"
 require "trailblazer/operation/model"
 require "trailblazer/operation/contract"
 require "trailblazer/operation/representer"